[database] 修改 product  `Discount` 型別為 int

+++ Database/shopdb_product.sql
@@ -30,7 +30,7 @@ CREATE TABLE `product` (
   `Describe` varchar(256) NOT NULL,
   `Quantity` int NOT NULL,
   `Price` int NOT NULL,
-  `Discount` decimal(3,2) NOT NULL,
+  `Discount` int NOT NULL,
   `Discount_period` datetime DEFAULT NULL,
   PRIMARY KEY (`PID`),
   KEY `Manager_ID` (`Manager_ID`),
@@ -45,8 +45,8 @@ CREATE TABLE `product` (
 LOCK TABLES `product` WRITE;
 /*!40000 ALTER TABLE `product` DISABLE KEYS */;
 INSERT INTO `product` VALUES 
-(1,_binary 'image',1,'block','This is a square',5,1200,1.00,'2023-12-12 12:59:23'), 
-(2,_binary 'image',1,'red','This is a book',9,300,100.00,'0000-00-00 00:00:00');
+(1,_binary 'image',1,'block','This is a square',5,1200,1,'2023-12-12 12:59:23'), 
+(2,_binary 'image',1,'red','This is a book',9,300,100,'0000-00-00 00:00:00');
 /*!40000 ALTER TABLE `product` ENABLE KEYS */;
 UNLOCK TABLES;
 /*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
